Restructured files and filenames. This is to better fit what I was making, as I was just going off an idea.
